requirements.txt
requests>=2.28.0
pyserial>=3.5
syrics>=0.1.0